code/pyproject.toml
[project]
name = "rlhf-book-code"
version = "0.1.0"
description = "Educational code examples for RLHF Book (https://rlhfbook.com)"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Core ML
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "datasets>=2.19.0",
    "accelerate>=0.30.0",

    # LoRA / quantization (for reward models)
    "peft>=0.11.0",
    "bitsandbytes>=0.43.0",

    # Config & logging
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "wandb>=0.17.0",
    "rich>=13.0.0",
    "tqdm>=4.66.0",

    # Policy gradients (procedural datasets)
    "reasoning-gym>=0.1.24",

    # Flash attention - default for x86_64, provides major speedup
    # Excluded on aarch64 (DGX Spark) where SDPA is used instead
    "flash-attn>=2.5.0; platform_machine == 'x86_64'",
]

[project.optional-dependencies]
# For DGX Spark (aarch64 / Blackwell) - no flash-attn, uses PyTorch SDPA
dgx-spark = []

# Development dependencies
dev = [
    "ruff>=0.4.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
train-policy = "policy_gradients.train:main_cli"
train-orm = "reward_models.train_orm:main"
train-prm = "reward_models.train_prm:main"

[tool.uv]
package = true

# PyTorch CUDA wheels - only use cu121 on x86_64 Linux
# aarch64 (DGX Spark) uses system PyTorch or cu130 wheels
[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu121", marker = "platform_system == 'Linux' and platform_machine == 'x86_64'" },
    { index = "pytorch-cu130", marker = "platform_system == 'Linux' and platform_machine == 'aarch64'" },
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "C4"]
ignore = ["E501", "B007", "F841"]

[tool.ruff.lint.isort]
lines-after-imports = 2
